migrate main availability page
